Changed size from int to double

+++ Java/book-metadata-parser/src/main/java/donotmodify/Dimensions.java
@@ -1,8 +1,8 @@
 package donotmodify;
 
 public class Dimensions {
-    public int width;
-    public int height;
-    public int depth;
+    public double width;
+    public double height;
+    public double depth;
     public String unit; //expect only cm and inches
 }
